Add support for building pip packages.
Added config.py and stuff, so that packages can be built.
Considering report/ as a module, added __init__.py for it.
To avoid circular dependency with mordred, I added
report/config.py (from mordred) for now. Maybe we should
remove that at all, since the functionality of using a
mordred config file is already present in mordred.