Add LaTeX templates to the package.
For adding the LaTeX templates for the package, I moved its directory
under the report directory, with the name latex_template. There are
other ways of doing this, but after some reading, and some attempts,
this one seems to be the most convenient.
I also modified report.py to accomodate to this change, and to
find templates where they are installed now.
Minor additional change: now version is read from a _version.py
file, as in other GrimoireLab packages.