[tests] Add test base class
This code proposes a test base class in charge of loading enriched
data to an ElasticSearch index. In a nutshell, for every test
method of a child test class, the base class loads enriched items
and mappings from two files within the folder `tests/data` and
save them to an ES index. The files and index are identified
according to the values of the attributes `name` and `enrich_index`
defined in the child test class (e.g., git_commit and git_enrich).

Note that the content of the files can be obtained by dumping
mappings and documents from existing enriched indexes using
`elasticdump`. For instance:

```
elasticdump \
  --input=https://...@blue6.biterg.io/data/github_enrich_issue_180713 \
  --output=./github_issue_mappings.json \
  --type=mapping \
  --limit=1000
elasticdump \
  --input=https://...@blue6.biterg.io/data/github_enrich_issue_180713 \
  --output=./github_issue.json \
  --type=data \
  --limit=1000
```