[tests] Update test base class
This code proposes changes to the `setUp` and `tearDown` methods of the
test base class to `setUpClass` and `tearDownClass`. Since all the
tests under a Test<Class> class will use the same elasticsearch index
(which is inserted into ES by the setup methods) using `setUpClass`
is a better approach then using `setUp` because `setUpClass` is run
only once at the beginning of the tests and `setUp` is run before every
test is run. This will make the tests faster when the base class is
used for other tests too.