[MS2/metrics] Change field for timeseries aggregations in ClosedIssues
Earlier the ClosedIssues timseries were being calculated using
`grimoire_creation_date` as the date field. This will lead to errors
as to the issues that were closed in a particular month.

This code changes the date field from `grimoire_creation_date` to
`closed_at` which will appropriately give us the number of issues
that were closed at in a particular period of time

This code also updates the test data for the same.