[MS2/latex] Separate github_issues and github_prs data for Overview
This code splits the efficiency-github_issues.tex file into two:
efficiency-github_issues.tex and efficiency-github_prs.tex.

Earlier, if github_prs data source was not added to the report, then a latex error
was thrown because the latex template, for the over view section of the report,
expected a column name from a file which generated from the github_prs data source
but since the data source was not added, the file wasn't created and hence
there was no column name matching that latex template.

This code tries to fix that bug by splitting the template file into two parts,
one for each data source.