[MS2/tests] Set up module level test fixtures
Currently, at the beginning of running each test class,
an ES index is created for that class from json files stored in
data/indices dir.

These indices are setup so that data from these indices can be used
to test the functions related to that data source (git, github_issues)

Now, since these indices will be used to generate the report too,
this code proposes to add Module level fixtures to the tests which
will index the data related to testing into ES only once, when the tests
aer starting to run. And these module level fixtures will also remove the
indices once the tests are complete.

In this way, we won't have to index the data again and again and this
will save us time.