Use closing date on GitHub PRs when computing BMI metric
BMI metrics for ITS data sources (i.e Jira or GitHub issues)
use 'closed_at' on closed issues. For GitHub PRs, 'updated_at'
was being used. This value is invalid because there can be
activity not related with the work on that PR (i.e updates
by GitHub maintainers).

Moreover, there was a typo in the name of the field used to
calculate the closed PRs.