[CI] Migrate from Travis CI to GitHub Actions
Travis CI has brought a few changes in its plans
which is causing friction to the current workflow.

This commit removes the support of Travis CI and adds
the support of GitHub Actions as the new CI service.

The version of matplotlib is also pinned to '2.0.2'
as a hot-fix for fixing the build as it was failing
because of the below error.
```
ImportError: cannot import name 'iterable' from 'matplotlib.cbook'
```

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>